Oops. I was committing from the wrong box.

+++ server/asobj/Global.cpp
@@ -167,13 +167,14 @@ as_global_unescape(const fn_call& fn)
     return as_value(input.c_str());
 }
 
+// parseFloat (string)
 static as_value
 as_global_parsefloat(const fn_call& fn)
 {
     ASSERT_FN_ARGS_IS_1
 
     as_value rv;
-    float result;
+    double result;
     
     std::istringstream s(fn.arg(0).to_string());
     
@@ -187,6 +188,7 @@ as_global_parsefloat(const fn_call& fn)
     return rv;
 }
 
+// parseInt(string[, base])
 // The second argument, if supplied, is the base.
 // If none is supplied, we have to work out the 
 // base from the string. Decimal, octal and hexadecimal are
